requirements.txt
affine==2.4.0
attrs==24.2.0
bokeh==3.6.1
certifi==2024.8.30
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rasterio==1.4.2
six==1.16.0
tornado==6.4.2
tzdata==2024.2
xyzservices==2024.9.0
